+++ jsonargparse/__init__.py
@@ -72,4 +72,4 @@ __all__ += _util.__all__
 __all__ += _deprecated.__all__
 
 
-__version__ = "4.43.0"
+__version__ = "4.44.0"